libchip: Fix prototypes

+++ c/src/libchip/rtc/m48t08.h
@@ -40,45 +40,45 @@ extern rtc_fns m48t08_fns;
  */
 
 uint32_t   m48t08_get_register(     /* registers are at 1 byte boundaries */
-  uint32_t    ulCtrlPort,           /*   and accessed as bytes            */
+  uintptr_t   ulCtrlPort,           /*   and accessed as bytes            */
   uint8_t     ucRegNum
 );
 
 void  m48t08_set_register(
-  uint32_t    ulCtrlPort,
+  uintptr_t   ulCtrlPort,
   uint8_t     ucRegNum,
   uint32_t    ucData
 );
 
 uint32_t   m48t08_get_register_2(   /* registers are at 2 byte boundaries */
-  uint32_t    ulCtrlPort,           /*   and accessed as bytes            */
+  uintptr_t   ulCtrlPort,           /*   and accessed as bytes            */
   uint8_t     ucRegNum
 );
 
 void  m48t08_set_register_2(
-  uint32_t    ulCtrlPort,
+  uintptr_t   ulCtrlPort,
   uint8_t     ucRegNum,
   uint32_t    ucData
 );
 
 uint32_t   m48t08_get_register_4(   /* registers are at 4 byte boundaries */
-  uint32_t    ulCtrlPort,           /*   and accessed as bytes            */
+  uintptr_t   ulCtrlPort,           /*   and accessed as bytes            */
   uint8_t     ucRegNum
 );
 
 void  m48t08_set_register_4(
-  uint32_t    ulCtrlPort,
+  uintptr_t   ulCtrlPort,
   uint8_t     ucRegNum,
   uint32_t    ucData
 );
 
 uint32_t   m48t08_get_register_8(   /* registers are at 8 byte boundaries */
-  uint32_t    ulCtrlPort,           /*   and accessed as bytes            */
+  uintptr_t   ulCtrlPort,           /*   and accessed as bytes            */
   uint8_t     ucRegNum
 );
 
 void  m48t08_set_register_8(
-  uint32_t    ulCtrlPort,
+  uintptr_t   ulCtrlPort,
   uint8_t     ucRegNum,
   uint32_t    ucData
 );
